todo-service-main/src/api/pyproject.toml
[project]
name = "todo-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Vyacheslav Atkin", email = "avea22r225@student.bmstu.ru"},
]
dependencies = ["fastapi>=0.115.6", "uvicorn>=0.32.1", "pydantic>=2.10.3", "sqlmodel>=0.0.22", "psycopg2-binary>=2.9.10", "dependency-injector>=4.43.0", "python-keycloak>=4.7.3", "pydantic-settings>=2.6.1", "alembic>=1.14.0"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
